Move MacAddr helpers.
