+++ nexus/db-model/src/macaddr.rs
@@ -8,70 +8,10 @@ use diesel::pg::Pg;
 use diesel::serialize::{self, ToSql};
 use diesel::sql_types;
 use omicron_common::api::external;
-use rand::thread_rng;
-use rand::Rng;
 
 #[derive(Clone, Copy, Debug, PartialEq, AsExpression, FromSqlRow)]
 #[diesel(sql_type = sql_types::BigInt)]
 pub struct MacAddr(pub external::MacAddr);
-
-impl MacAddr {
-    // Guest MAC addresses begin with the Oxide OUI A8:40:25. Further, guest
-    // address are constrained to be in the virtual address range
-    // A8:40:25:F_:__:__. Even further, the range F0:00:00 - FE:FF:FF is
-    // reserved for customer-visible addresses (FF:00:00-FF:FF:FF is for
-    // system MAC addresses). See RFD 174 for the discussion of the virtual
-    // range, and
-    // https://github.com/oxidecomputer/omicron/pull/955#discussion_r856432498
-    // for an initial discussion of the customer/system address range split.
-    // The system range is further split between FF:00:00-FF:7F:FF for
-    // fixed addresses (e.g., the OPTE virtual gateway MAC) and
-    // FF:80:00-FF:FF:FF for dynamically allocated addresses (e.g., service
-    // vNICs).
-    //
-    // F0:00:00 - FF:FF:FF    Oxide Virtual Address Range
-    //     F0:00:00 - FE:FF:FF    Guest Addresses
-    //     FF:00:00 - FF:FF:FF    System Addresses
-    //         FF:00:00 - FF:7F:FF    Reserved Addresses
-    //         FF:80:00 - FF:FF:FF    Runtime allocatable
-    pub const MIN_GUEST_ADDR: i64 = 0xA8_40_25_F0_00_00;
-    pub const MAX_GUEST_ADDR: i64 = 0xA8_40_25_FE_FF_FF;
-    pub const MIN_SYSTEM_ADDR: i64 = 0xA8_40_25_FF_00_00;
-    pub const MAX_SYSTEM_RESV: i64 = 0xA8_40_25_FF_7F_FF;
-    pub const MAX_SYSTEM_ADDR: i64 = 0xA8_40_25_FF_FF_FF;
-
-    /// Generate a random MAC address for a guest network interface
-    pub fn random_guest() -> Self {
-        let value =
-            thread_rng().gen_range(Self::MIN_GUEST_ADDR..=Self::MAX_GUEST_ADDR);
-        Self::from_i64(value)
-    }
-
-    /// Generate a random MAC address in the system address range
-    pub fn random_system() -> Self {
-        let value = thread_rng()
-            .gen_range((Self::MAX_SYSTEM_RESV + 1)..=Self::MAX_SYSTEM_ADDR);
-        Self::from_i64(value)
-    }
-
-    /// Construct a MAC address from its i64 big-endian byte representation.
-    // NOTE: This is the representation used in the database.
-    pub(crate) fn from_i64(value: i64) -> Self {
-        let bytes = value.to_be_bytes();
-        Self(external::MacAddr(macaddr::MacAddr6::new(
-            bytes[2], bytes[3], bytes[4], bytes[5], bytes[6], bytes[7],
-        )))
-    }
-
-    /// Convert a MAC address to its i64 big-endian byte representation
-    // NOTE: This is the representation used in the database.
-    pub fn to_i64(self) -> i64 {
-        let bytes = self.0.as_bytes();
-        i64::from_be_bytes([
-            0, 0, bytes[0], bytes[1], bytes[2], bytes[3], bytes[4], bytes[5],
-        ])
-    }
-}
 
 NewtypeFrom! { () pub struct MacAddr(external::MacAddr); }
 NewtypeDeref! { () pub struct MacAddr(external::MacAddr); }
@@ -95,20 +35,6 @@ where
 {
     fn from_sql(bytes: RawValue<DB>) -> deserialize::Result<Self> {
         let value = i64::from_sql(bytes)?;
-        Ok(MacAddr::from_i64(value))
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::MacAddr;
-
-    #[test]
-    fn test_mac_to_int_conversions() {
-        let original: i64 = 0xa8_40_25_ff_00_01;
-        let mac = MacAddr::from_i64(original);
-        assert_eq!(mac.0.as_bytes(), &[0xa8, 0x40, 0x25, 0xff, 0x00, 0x01]);
-        let conv = mac.to_i64();
-        assert_eq!(original, conv);
+        Ok(MacAddr(external::MacAddr::from_i64(value)))
     }
 }
